feat: add documentation regarding using Job Bundles as an integration
Summary:
Job Bundles are a useful way to define Jobs to submit to Amazon Deadline Cloud,
but there is no documentation on what they are or how to construct and use one.
This commit aims to remedy that by adding a README to the job_bundles directory
that outlines the components of a Job Bundle, and the applications that
Deadline Cloud has for submitting Jobs with one.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>