feat: adds link to job template guide and info about job env vars
Summary:

The Open Job Description wiki now has a guide covering an introduction to job templates
that walks the reader through the creation of a job template, and introduces them to concepts
along the way. This adds a breadcrumb to that guide in the section on job templates.

We also add information on the `DEADLINE_*` environment variables that are available to a Job.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>