docs: enhance contributing guidelines
Problem:

The contributing guidelines in this repository assume a fairly high level of familiarity with GitHub
workflows and mostly serve as memory jogs for  developers. This makes it challenging for those who
are not familiar with our workflows to understand how to go about making a contribution.

Solution:

Rework/refactor the CONTRIBUTING.md, GitHub issue templates, and pull request template with the
goal of improving understandability to those not familiar with the way that we use GitHub pull request
workflows, and our use of conventional commits.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>