refactor: Apply fix in anticipation of deadline-cloud warning update
In https://github.com/aws-deadline/deadline-cloud/pull/673, we are
changing some behaviors of the job attachments upload warnings. One
detail is that the parameter_values.yaml/json file is not treated as
known paths, while the parameter values passed directly to the function
call are.

This change passes the parameter values to create_job_from_job_bundle,
so that it will continue to submit successfully after that
deadline-cloud PR is merged.

Signed-off-by: Mark Wiebe <399551+mwiebe@users.noreply.github.com>